+++ src/GeoCookie.php
@@ -10,16 +10,18 @@
 
 namespace lukeyouell\geocookie;
 
-use lukeyouell\geocookie\services\GeoCookieService as GeoCookieServiceService;
-use lukeyouell\geocookie\variables\GeoCookieVariable;
 use lukeyouell\geocookie\models\Settings;
 use lukeyouell\geocookie\twigextensions\GeoCookieTwigExtension;
+use lukeyouell\geocookie\utilities\Log;
+use lukeyouell\geocookie\variables\GeoCookieVariable;
 
 use Craft;
 use craft\base\Plugin;
-use craft\services\Plugins;
 use craft\events\PluginEvent;
+use craft\events\RegisterComponentTypesEvent;
 use craft\helpers\UrlHelper;
+use craft\services\Plugins;
+use craft\services\Utilities;
 use craft\web\Request;
 use craft\web\twig\variables\CraftVariable;
 
@@ -30,33 +32,22 @@ class GeoCookie extends Plugin
     // Static Properties
     // =========================================================================
 
-    /**
-     * @var GeoCookie
-     */
     public static $plugin;
+
+    // Public Properties
+    // =========================================================================
+
+    public $schemaVersion = '1.1.0';
 
     // Public Methods
     // =========================================================================
 
-    /**
-     * @inheritdoc
-     */
     public function init()
     {
         parent::init();
         self::$plugin = $this;
 
         Craft::$app->view->registerTwigExtension(new GeoCookieTwigExtension());
-
-        Event::on(
-            CraftVariable::class,
-            CraftVariable::EVENT_INIT,
-            function (Event $event) {
-                /** @var CraftVariable $variable */
-                $variable = $event->sender;
-                $variable->set('geoCookie', GeoCookieVariable::class);
-            }
-        );
 
         Event::on(
             Plugins::class,
@@ -67,6 +58,20 @@ class GeoCookie extends Plugin
                 }
             }
         );
+
+        Event::on(
+            Utilities::class,
+            Utilities::EVENT_REGISTER_UTILITY_TYPES,
+            function (RegisterComponentTypesEvent $event) {
+                $event->types[] = Log::class;
+            }
+        );
+
+        // Register components
+        $this->setComponents([
+            'geoService' => \lukeyouell\geocookie\services\GeoService::class,
+            'logService' => \lukeyouell\geocookie\services\LogService::class,
+        ]);
     }
 
     // Protected Methods